lambda/requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
Faker==22.0.0

